requirements.txt
streamlit==1.28.1
tensorflow>=2.10.0,<2.16.0
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
Pillow==10.0.0
opencv-python==4.8.0.76
scikit-fuzzy==0.4.2
deap==1.3.3
kaggle==1.5.16
